Removes link_directory in favour of find_library
The official documentation of CMake advises against using the
`link_directory` function and instead advises you to use the
`find_library` function.

This commit changes to using `find_library`. The believed side effect
of this is that our binary may now figure out to use inbuilt libraries
should they not exist in the dependencies folder.