Rewrites build.sh. Removes set compiler from CMake
The build.sh file has been rewritten to be more robust and support more
options. In addition, the CMakeLists no longer directly sets the
compiler and instead depends on CC and CXX environmental flags.

However, since this project does not yet compile successfully with
clang++, we had to make a temporary compromise. `build.sh` will now
store the current environmental variables of CC and CXX and export gcc
and g++ as the compiler of choice until the compilation is completed.

Since it only sets it temporarily, it should have no effect on your
environmental variables. As soon as clang++ can compile the project, we
should remove this functionality from the file