Fixes some of the warnings that GCC now pumps out
Lots and lots of warnings after I added the extra warnings,
most of them are limited to the math library. Since we'll replace the
math library with MMM made by @Thhethssmuz, we'll ignore them for now.

We'll discuss their viability when we've implemented MMM as a
dependency of this project