CMake no longer globs all our files
As a larger change, CMake no longer globs all our files and they are
instead listed directly within the CMakeLists.txt.

There's a downside and an upside to this change.

The downside is that you'll need to add the hpp and cpp files
whenever you're adding new files. You will never need to add TPP files.

There's two upsides, actually. The first one being that you no longer
need to run the `cmake` command when you're adding file as you're
actually having to change the CMakeLists.txt and the makefile detects
that.

The other upside is that when you pull changes from the repo, you'll
always know that you can just run `make`, since all new files have to be
added to the CMakeLists therefore changing it, which, again, the
makefile detects automatically.

Fixes #1