Adds newline between includes in OpenGLHeaders
This is so that clang-format does not reorder the includes. It is very
important, crucial in fact, that the OGL header comes before the GLFW
header. The whole reason for the OpenGLHeaders.hpp file is to make sure
that this always happens.

FYI, including GLFW before OGL will cause a compile time error where
OGL complains about GL headers included prior to it being included.