Removes cotire
The performance difference between using cotire and not using cotire
where so insignificant that there's no point for us to use it. We
barely utilize header-only library and those that do take time to
compile are not header-only ones.

By just testing quickly, there was a 0.09s difference between using
cotire and not using cotire for our sake.