Disable several dependency options in CMake
A lot of the dependencies were building their tests, examples and what
not. This commit changes our CMakeLists to set most of these options to
false. This reduces the files it has to build from ~1000 to ~500.

We probably won't ever need them. If we do though, they're just a
boolean change away.