Adds new handler for char callback
GLFW supports a char callback, which is being called with the text that
is inputted. This means that we wont have to do conversions from a key
character to a text character.

It does however only give us the code to the unicode character, so that
does have to be turned into a string if it is to be showed on the screen