Moves Input to its own namespace together with Event
Due to this, I also also moved some of the variables, which really only
existed within Input class due to the namespace it gave. Yes, we now
have Input::Input, but that's okay.

The Event class has been rewritten to support all types of events and
has a type which indicates the event type, for instance it can have the
type Input::Event::Type::MousePressed. Explicit.