Event system now works properly
The newly implemented event system now works as it should. It will no
longer depend on the return type of a function, but instead requires
handler functions to call stopPropagation once they've done something.

There are still some few glitches to work out, but it works pretty much
as it should.