Now uses a ResourceManager for all resources
In order to make sure that this was properly implemented, a lot of files
had to be changed. For instance, Texture is no longer a drawable
element. The same features of that of the previous Texture can be
implemented by using GL::Rectangle and telling it to use a texture.

The following are now considered resources and are loaded according
to which state/scope they're required in:

  - Texture
  - Font
  - Program

The list may be expanded as we see fit. All these classes have a base
class, `Resource`, in common. This base class stores information such
as filename, whether it is loaded or not and what scope/state it belongs
to.

As said, since this required a larger overhaul of all the resources
used, a lot of changes where made. Luckily, the changes were mostly
limited to the GUI elements, as these are almost the only thing in the
engine that uses these resources.

Lastly, the ResourceManager loads a resource file in media. This file
specifies, in LUA, what resources to load and when to load them.