Lua folder now duplicates the rest of the project
Since the Lua definitions will be extended with almost everything in the
engine, I thought it would be smart to make sure it had some structure
that simulated the structure that we already had within the project to
make it easier to see what each file represents.

It should be noted that none of the files under the Lua folder (except
for Lua.hpp) should ever be included by any other file than Lua.cpp.