Fixes several issues related to resources
When settings are changed, the current state or the whole engine may
require a reload for some specific settings to be applied.

For instance, changing the resolution requires a state reload since so
many of the elements require the resolution to set themselves. Another
case is if the multisampling of the application is changed. This change
requires the OpenGL context to be recreated.

When performing these reloads, some of the resources were not
loaded/unloaded properly. This could cause some undefined behaviour,
such as text not rendering.

This commit fixes these issues by thoroughly checking whether or not the
resources are loaded/unloaded. In addition, it also moves window
refreshes to the last thing to do in the loop.

The latter change makes sure that GLFW has completed all that it should
do and will not require the window until the start of the next loop.