Adds foreground and background colours to text
The text can now use a special syntax to set colours of words/characters.
The colours use 0-255 RGBA colour definition.

In order to not break any other text that is inserted, you have to
explicitly escape the starting character.

The syntax of colours is as follows:

  \<Fr,Fg,Fb,Fa:Br,Bg,Bb,Ba>\</>

where the F is `foreground` and B is `background`, r is `red`, g is
`green`, b is `blue` and a is `alpha`.

Each of the colour values are separated by ',' and the background and
foreground colours are separated by ':'.

For instance, to colour the word `I have colours` in three different
colours:

\<255,0,0,255:>I \<0,255,0,255:>have \<0,0,255,255:>colours\</>\</>\</>

The above colours `I` in red, `have` in green, `colours` in blue. It does
not include background colours and uses the default for that.