Adds more detailed information about dependencies
This commit adds more information about each and everyone of our
dependencies, linking to their repository or main page where applicable.
In addition, mentions and links to all the differences licenses has also
been included.

By doing this, I made sure to check that we did not use any library that
would be incompatible with a `MIT` license (or any other license for
that matter). As long as we compile MultiNeat as a shared library, we
should be able to license this project under whatever license we want.