Abstracts Drawable into two separate classes
Since the Drawable class can eventually be used by the GUI elements and
we still want to use it for 3D elements. Due to this, we now have a new
class, Drawable3D, that contains all things that are in common for all
the 3D elements, in particular, physics elements.