Adds possibility of loading debug texture
If a texture that is requested does not exist, it will use the debug
texture. This won't make the engine crash, but it'll still give a good
indication of what is wrong. It will also log an error to the console
saying that the requested texture does not exist.

There is currently not a debug texture in the resource file, but this
will soon be fixed.

**Note**: It will still throw error if the Texture you request exist,
but is not available for your scope.