Restructures Lua code into separate folders
The Lua code is now structured in a more sensible way. For instance, the
`console` folder now contains all the code that is meant for and used
with the console. `gui` contains all the gui elements and `debug`
contains code that is meant to help debugging

This structure easily allows for conditional loading of modules. For
instance, adding a `console.enabled` to the CFG file and using this in
Lua would allow you to disable console features