Abstracts the Library loading a bit more
To allow for us to have more than one instance of the Lua class, I've
abstracted the code a bit so that we can specify which libraries we want
to load when creating the instance.

The new instance will save the list of libraries that are to be loaded
and use these each time `reInitialize` is called. This makes sure that
the class has exactly the same features as before.

When the time comes, we can possibly make an API that can be used by
someone else to program something for our game by simply saying which
libraries we want to expose to them.