Rewrites mesh resource to better support names
The mesh resource will no longer structure its mesh elements in a tree,
but instead store these in a vector. The new changes also multiplies out
the transformations, so less multiplication is done when the meshes
should be drawn.

In addition, the mesh class can now be queried for specific submeshes by
both index and a string name, where the name is applied to the submesh
using a 3D editor such as blender