Adds MeshPart which should work with Meshes
Often you have larger meshes that have lots of submeshes, such as the
case with our spider. In order to allow the submeshes to have physics
and to be drawn with the positions outputted by our physics engine, we
need to have these as their seperate Drawable3D classes.

This class makes sure that this is possible.