Adds functions to copy and delete objects
Since the btRigidBody elements should exist one for each instance within
the game, they have to be duplicated based on the information in the
original import.

In addition to this, the constraints also have to be copied and set to
use the newly constructed btRigidBody elements instead of the original
imports.

To simplify this process, there now exists two new functions that does
the copying and deleting of these objects.