Uses new material structure for multiple textures
Since a mesh can have multiple textures, which are divided into
materials, the loader should be able to load and use these properly.

When Assimp loads a mesh, it checks if that mesh has multiple materials.
If it has multiple materials, it will split the mesh up until there are
multiple meshes with only one material each.

In order to handle this, the loader now loads meshes and its material to
a separate structure. This structure contains the startIndex and number
of vertices for its mesh, while the SubMesh class contains the
startIndex and number of vertices for all its materials/submeshes
combined.

The benefit of this is that it allows meshes to have multiple materials,
but also allows one-draw call per mesh when these materials arent
needed, such as in the case of shadowmapping