Adds new world per spider as it performs better
There were significant differences between having one world per spider
or adding all spiders to the same world and ignoring the collisions
between each spider.

The problem was that even if the collisions were ignored, each RigidBody
still had to be checked for whether or not collision detection should
execute.