Adds options of solvers and phases to World
The type of solver and phase can now be chosen through enums within the
World class. Selecting one is as simple as changing the argument when
creating a World object.

As far as testing goes, Dantzig seems to be the better out of the MLCP
solvers but does not compare to a Standard solver with increased
iterations when it comes to performance.