Adds storing of best and logs breakdown of it
In addition to performing the final calculations on the fitness and
logging the best one like usual, it also logs the statistics of the
best one so that we can see where it excels.

In addition, the generations are now stored as a variable on the spider
swarm (may let us run multiple spiderswarms if we want) and stores
the best one of all generations just to see how all the other
generations are doing.