Splits the Model.fs file into lib files
Since the Model.fs file has a lot of the code that will be used in other
fragment shaders, I've split it up and used the include statement to add
them together.

The only issue is that if any errors occur, the line numbers are a bit
off. This has been solved by priting the entire shader in its full glory
whenever an error happens, with linenumbers, so that you can see where
the error is.

The files that are included should be recognizable since they all have
their own module comment.

The file start off with declaration and documentation statements so that
you dont need to look at the implementation of the functions to
understand what the file includes.