Adds seperate fitness class
Since the fitness calculation was all over the place and not really that
easy to document / follow, I've added a seperate class that holds the
following:

- Short name of the fitness calculation
- Longer description
- A calculation function
- A finalize function

The shorter name should be short enough to be logged and the longer
description should give a detailed explanation of what it tries to
achieve.

The calculation function is a function that is executed after each
update, giving you the option of returning a fitness.

The finalize function is used at the end of a simulation to set
the final fitness values.