Allows closing of console using same opening key
This change allows you to both open and close the console with the same
key. This has one disadvantage that you wont be able to write the letter
that the key is bound to. It is therefore recommended to attach it to a
key that you either never use or one that does not have a visual
representation, such as the F-keys.