Fixes segmentation fault on exit
When the current state was deleted, everything it used of resources was
already deallocated. Normally, this would be entirely fine as nothing
that used these resources really required the resources to delete
themselves.

However, Bullet required some of the resources in PhysicsMesh
to delete itself. When these pointers were nullpointers, it would
crash when trying to use them.

@Thhethssmuz, I fixed the infamous segmentation fault. Yay.