Removes experiment results due to size
First off, these will be added back, but with less duplication.
Currently, there is a lot of duplication going on here. For every new
best, there's a CSV, population and substrate file that are duplicate.
Only the final CSV is needed, as that records everything before it.
Stripping this from the repo and removing it from the history in order
to save space as git loves to keep everything in its .git folder.